types: rename api structs
Summary: Given that the on-the-wire representation of file and tree data is essentially the same, we can reuse our existing structs for file fetching for tree fetching. As such, we should rename these structs to make it clear that they are not just for file data. (This also involves updating some comments/variable names to make sure everything is consistent.)

Reviewed By: singhsrb

Differential Revision: D15286690

fbshipit-source-id: 8c7fa4392057e90a9f19beb17b0bbcbf04b7e8e7