edenapi: make sure EdenApi trait is object safe
Summary:
After removing the Hyper based EdenAPI client, there is at present only one client implementation. As such, the Python bindings have been updated to use that implementation directly rather than a `Box<dyn EdenApi>` trait object.

In spite of this, it would be nice to ensure that this trait remains object safe so that if we add a new implementation in the future, we can easily go back to using a trait object. (Otherwise, it is very tempting to add `impl Trait` or generic methods to the trait which would be difficult to remove later.)

This diff adds a simple compile-time check that ensures the trait is object safe.

Reviewed By: quark-zju

Differential Revision: D15289514

fbshipit-source-id: e880f22fa7f48b52467aedf41036711a1713deb9