ui: write to stderr in a single operation
Summary:
Writes to stderr may be unbuffered.  This means when `ui.write_err` is called
with multiple strings, they may end up being separated as different writes.
This causes flakiness in the test, particularly when receiving error messages
from a remote server, as depending on timings output may appear as either:

    remote: abort: error message

or

    remote: abort:
    remote: error message

Write to the error stream in a single operation by joining all the messages together.

Differential Revision: D15316694

fbshipit-source-id: de5edf05e137d88006a0919f02f8b9cf7a17fa42