help: drop internal help from UI
Summary:
Drop the "internals" help text from the UI. This is because the internal docs
are mostly outdated (ex. we're moving to new HTTP protocols and slowly removing
bundle2 formats, we're moving away from revlogs).

This diff just disabled access to them via the `hg help` command, the files are
still accessible in the repo.

Reviewed By: kulshrax

Differential Revision: D15291675

fbshipit-source-id: a6e55e0ed085bb54def3790aaaf1a92a4bfa787f