mononoke: implement bundle combination boilerplate code
Summary: This diff adds some boilerplate code paving the way to the "merge bundles" idea (see attached task). Note that the way it is implemented now it never mreges anything and even if someone tries to pass `--combine-bundles` with >1 value, it will panic. This is done deliberately, because there's still work to do: mainly on the Mercurial side, but also around properly asynchronous timestamp file merging. I want to land this code, because it gets the correct types in place, which is the hardest thing to incrementally fix going forward.

Reviewed By: krallin

Differential Revision: D15245375

fbshipit-source-id: e476d6c68989a83adeb1ade8ccf1aec91fc9df24