pathmatcher: add a matcher that is tree aware
Summary:
We're in need of some "protocol" to select files by patterns. For example, user
should be able to request "me all trees / files related to fbcode/scm/hg directory",
or complex patterns like "everything in fbcode/, excluding fbcode/experimental,
but do include fbcode/experimental/**/TARGETS".

The pattern of choice is quite similar to gitignore rules, with some
differences to make it harder to create inefficient matchers stated in
docstring.

To make the matcher efficient, it needs to be able to answer fast path
questions about parent directories. For example, if "/fbcode*/scm/hg/**" is the
only pattern, the matcher should be able to answer questions in this way:

  - should visit "fbobjc"? no! skip it entirely
  - should visit "fbcode"? yes! but not everything
  - should visit "fbcode/folly"? no! skip it entirely
  - should visit "fbcode/scm/hg"? yes! and everything inside it without further
    testing

The TreeMatcher added by this diff sastifies the above needs.

In the future, it could potentially be used to replace the sparse profile
matcher, or potentially clean up the Python matcher.py.

Reviewed By: DurhamG

Differential Revision: D13502750

fbshipit-source-id: e3fdf9e61a34508b57b1c12f0be4e7ad5089e72c