treematcher: provide a `expand_curly_brackets` function
Summary:
There are 2 problems this function helps solving:

1. `globset` used by TreeMatcher does not support nested round brackets like
   `{a,{b,c}}`
2. TreeMatcher cannot handle patterns like `{a,b/c}` efficiently because `/` is
   in it.

Once the brackets are removed, it's possible to have more fast paths in
TreeMatcher.

 ---

TreeMatcher was initially the Rust version of `mercurial.matcher._tree`, it was
sent as D10861610 but the code was changed greatly since then. So I reordered
commits and TreeMatcher in D10861610 is no longer a thing at this commit.
TreeMatcher referred above points to the upcoming gitignore matcher that knows
how to skip directories efficiently.

I lately found out that the csh style `{}` patterns are not part of the
official gitignore format, but got some support from the Rust `globset` crate.
So this utility function becomes less useful. It still looks interesting
though, therefore this commit.

Reviewed By: DurhamG

Differential Revision: D13502749

fbshipit-source-id: d2abd26a68963ff804a3c58c91c6c5c7e26864ed