edenapi: make MultiDriver take a callback to handle results
Summary:
Previously `MultiDriver` (a struct with manages a `curl::Multi` session) would collect all of the received HTTP responses and return them after all transfers had completed. This was problematic in situations where the number of requests is large; the process can run out of memory buffering this data. A better approach would be to deal with the data the moment it is received.

This diff changes `MultiDriver` to take a callback which can handle each response as each individual transfer completes, rather than waiting until the end. For now, the EdenAPI client simply sets a callback that adds the data to a Vec, meaning that we're still buffering. In a later diff, we'll update the callback to directly write the data to storage instead.

Differential Revision: D15379276

fbshipit-source-id: ab4f98e49141f6582b6e2d366045eda514404197