manifest: add finalize method
Summary:
The plan is to have finalize do the complicated history managing stuff and
have flush be a light weight content only hashing approach to storage.
Flush will be used in tests in the short term.
Finalize would be used in Python to write to Revlogs and handle linknodes.

(Note: this ignores all push blocking failures!)

Reviewed By: quark-zju

Differential Revision: D15290171

fbshipit-source-id: 54ed5c791254507fd9f2e874315284b70bc6b597