Initial blobrepo checker
Summary:
When underlying storage layers give us trouble, we want to be able to validate that nothing is broken.

Use `BlobRepo` to fetch all the data for history of a given bookmark, so that we can be confident that we have sufficient data to not lose history.

Later diffs will add support for checking the saved Mercurial changesets, too

Reviewed By: StanislavGlebik

Differential Revision: D15084862

fbshipit-source-id: 8c7184cf1cd0d52ca2ca9c8ce1d7e97b1e51db6d