commitcloud: add test demonstrating bad sync behaviour when removing obsolete bookmark
Summary:
If the user has a bookmark pointing at an obsolete commit, and then moves that
bookmark away from that commit, commit cloud can't tell whether the commit
should really be obsolete or not.  Previously the bookmark was the only thing
keeping the commit visible, but commit cloud doesn't know that.

I can't think of a way to fix this with obsmarkers.  It works with separate
mutation and visibility (T34204947).

Reviewed By: quark-zju

Differential Revision: D15097845

fbshipit-source-id: 5926553b4878b28823a6ce47d29c0346cba04e5e