revisionstore: add flush to MutableDeltaStore
Summary:
A MutableDeltaStore isn't object safe due to the close method. This means that
we can't use them in generic context and the code has to work around this
limitation.

Let's replace the close method by a flush one, that no longer takes the
ownership of self. The behavior of flush is to simply make the previous writes
visible, while still allowing future writes to the same object.

Reviewed By: kulshrax

Differential Revision: D15416709

fbshipit-source-id: 11c96d672171eaac91add3d7bcad7f8eae70de16