remove the hg store tester program
Summary:
This utility was initially used for testing flatmanifest and treemanifest
import.  It's less important now that we are planning to drop flatmanifest
import.

The recent changes to this code have mostly just been keeping up with changes
to Eden's initialization code.  I plan to continue making some additional
changes to Eden's config and local store initialization.  Rather than
continuing to keep this code up-to-date it seems simpler to just delete it.
We can always resurrect this later if we find a need for it again.

Reviewed By: chadaustin, strager

Differential Revision: D15456687

fbshipit-source-id: dd65f93012530392b399a353f5b092192666f102