revlog: avoid validating the hash if we just computed it
Summary:
In some cases the revlog actually computes the hash it uses. In that
situation we don't need to validate it. This is showing up as a performance
hotspot in profiles of the server receiving lots of commits. I'm not sure why
the creation of the hash is not showing up as a hotspot though, which is a bit
strange.

Reviewed By: kulshrax

Differential Revision: D15459193

fbshipit-source-id: 99823dd9e1709dbfecf2626b07f0e4c3ec43686c