remotefilelog: use 'HTTPS' instead of 'HTTP' in user-facing output
Summary: On the Source Control team, we've been referring to the project to replace parts of the Mercurial wire protocol with the Mononoke API server as "HTTP data fetching". This has caused confusion for other teams in the past (particularly teams with a security-related focus), as this nomenclature makes it seem as if we're using unsecured HTTP. (This is not true; we are always using HTTPS.) In light of this, let's make sure that user-facing messages always use the term "HTTPS" instead of "HTTP" to avoid this confusion.

Reviewed By: quark-zju

Differential Revision: D15465390

fbshipit-source-id: b556ed222dc01979ceaa0d4a7aa921cb1c38d75e