minibench: abstract what and how to measure
Summary:
The `minibench` library was initially created to measure the "minimal"
"floating numbers", runs functions "40 times", and assumes the result floating
number means "seconds". Later it got a "bench_once" function for calling a
function just "once", and assumes the result is an "integer", and means "bytes".

This diff defines a `Measure` trait to abstract away the "minimal",
"floating numbers" / "integers", "40 times" / "once", "seconds" / "bytes".

With the change, `bench` becomes the only top-level function, `min_run` and
`bench_once` are removed.

The existing wall clock and bytes measurement was re-implemented under the
`Measure` trait. They are slightly improved in output formatting, and run time
control (if a function takes too long to run, WallClock will not run it for 40
times).

The `elapsed` function is kept as an alias of `WallClock::measure` so most
existing users of minibench don't need change.

Reviewed By: xavierd

Differential Revision: D15454967

fbshipit-source-id: 22ee8ba330c4588ce01ca34a40c6efad5e650a27