indexedlog: add benchmark about Log::sync with indexes
Summary:
There are some "easy" optimizations to the `Log::sync` operation.
To make sure the optimizations are valid, benchmarks are added first.

The `Log::sync` operation is complex because it does both read and write.
Either code path will drop the in-memory indexes and rebuild them, which
makes it expensive.

Add a comprehensive set of benchmarks to measure it.

A sample output is:

  log sync (no index)                                 1.846 ms     680 B/212.993MB
  log sync (write)                                  108.103 ms   1.760KB/370.366MB
  log sync (read, no-op)                              0.200 ms   2.960KB/      0 B
  log sync (read, on-disk index lag)                 29.901 ms     840 B/      0 B
  log sync (read, log lag)                            0.168 ms   2.960KB/      0 B
  log sync (read 1, write N)                        108.138 ms   1.760KB/370.364MB
  log sync (read N, write 1)                          0.888 ms   3.000KB/ 16.680KB

Reviewed By: xavierd

Differential Revision: D15446784

fbshipit-source-id: 558e34c8b4e1ecb6f9858b843c00c9e9cecb21ad