edenapi: fix panic when downloaded size is 0
Summary: If the downloaded size is 0, we were crashing due to an array out of bounds access because we were calculating an array index with `f64::log10()`, which returns negative infinity (which is then cast to `usize::MAX`). Solve this by checking if the operand is 0. Given that the input is an unsized integer, 0 is the only special case we need to handle.

Reviewed By: xavierd

Differential Revision: D15471421

fbshipit-source-id: 226614a7f070f2d5a035058d9a460aa8869c6785