revlog: implement more efficient not-contains check
Summary:
When writing to a revlog we need to check if the revlog already
contains the node. A normal `node in self.nodemap` is quite expensive in large
repositories because it's likely the node is not in the nodemap and therefore it
must scan the entire index.

This new function uses the knowledge about the parents to limit how many nodes
are inspected when adding a new revision. In the degenerate case this could end
up traversing a large chunk of the graph in a slow way, so we catch the case
where the parents are far from the tip and revert to the old native path.

Reviewed By: quark-zju

Differential Revision: D15444548

fbshipit-source-id: 5f19ccf85fb11f1c787ebda09df9c4bef17a2a31