bindings: add mutabledeltastore python binding
Summary:
While the datapack read path is currently being done in Rust, the write path is
still being performed in Python. This class is the first step towards switching
the write path to Rust.

A nice bonus of this class is that it abstract packfile/indexedlog and besides
at construction time, the users of it won't be aware of which store is used.

Reviewed By: kulshrax

Differential Revision: D15474313

fbshipit-source-id: 49454d8f089089f164d858d590f254c78d3f5eb6