edenapi: perform network and disk I/O in parallel
Summary: Previously, we were alternating between asynchronously performing network I/O and synchronously blocking on disk I/O, which limited our end-to-end throughput for data fetching. This diff moves the network I/O to a separate thread where it can be handled asynchronously without interruptions. Received data is buffered in a channel until the main thread writes it to disk.

Reviewed By: xavierd

Differential Revision: D15476860

fbshipit-source-id: a601f8015b091bb49abbca0e68d91269e748a66e