edenapi: move multi_request out of EdenApiCurlClient
Summary: The `multi_request` (and `easy`) methods on `EdenApiCurlClient` don't really need to be methods; they only depend on `self` in one place respectively, which can easily be refactored into an argument. Removing the dependency on `self` makes it possible to use this method in closures that demand a `'static` lifetime, as is required by `thread::spawn` (used in the next diff in this stack).

Reviewed By: xavierd

Differential Revision: D15478200

fbshipit-source-id: 44a5d18652c5f33b198d60b446db0fdb42398bc0