manifest: bump version of once_cell to 0.2.0
Summary: A recent update to the crates in tp2 bumped the `once_cell` crate to version 0.2.0. This broke the build because the `Cargo.toml` for the `manifest` crate specified version 0.1.8. Apparently just changing the crate version to 0.2.0 fixes the build, so we weren't affected by whatever breaking changes were made to the crate.

Reviewed By: DurhamG

Differential Revision: D15492142

fbshipit-source-id: 552b0a751ab7c2aab5f0fbcb1124de4ea427790c