edenapi: don't send Option through channel
Summary: Previously, we were sending `Option<DataEntry>`'s through the channel from the I/O thread to the main thread so that `None` could be used to signal that all files had been downloaded. Turns out this isn't necessary in the current structure of the code, because the I/O thread will just drop the sender, causing the receiver to stop iteration.

Reviewed By: xavierd

Differential Revision: D15492366

fbshipit-source-id: 9ca883bd0afdb8f89909023ef0c4571e183c512a