edenapi: measure http request latency
Summary: Measure latency for HTTP requests, which in this case is defined as the time elapsed between the request being sent and the first byte of the response being received. The metric is rather crude, but I'm not sure of a better way to understand how long it takes before we get data from the server.

Reviewed By: xavierd

Differential Revision: D15493113

fbshipit-source-id: 642b61ee233436e664a46aee74e86814f8937394