mononoke: make sync job able to skip entries which were successfully applied
Summary:
In some cases, the sync binary gets killed after it had synced the bundle,
but before it had updated the counter in the database. This causes any
sunsequent sync to fail with the "conflicting changes" error. One of the ways
to mitigate this problem is to check on failure whether the bookmark on the server
is already pointing to the expected location, and pretend that the sync
succeeded, if it is.

Reviewed By: StanislavGlebik

Differential Revision: D15460956

fbshipit-source-id: 55d2df3545f8a34ebc61eb6da493b8d244629089