indexedlog: optimize no-op Log::sync path
Summary:
If there are nothing changed, do not reload anything. This makes `Log::sync` as
cheaper. An even cheaper version might be `stat`ing the metadata file first,
but that risks correctness issues, and `sync` is not designed to be called that
frequently anyway.

Before:

  log sync (read, no-op)                              0.215 ms   2.960 KB/      0 B
  log sync (read, on-disk index lag)                 32.489 ms     840 B /      0 B
  log sync (read, log lag)                            0.168 ms   2.960 KB/      0 B

After:

  log sync (read, no-op)                              0.031 ms   1.040 KB/      0 B
  log sync (read, on-disk index lag)                  0.031 ms     840 B /      0 B
  log sync (read, log lag)                            0.178 ms   4.000 KB/      0 B

Note: the "log sync (read, on-disk index lag)" is an extreme test case since we
wouldn't usually expect the index to lag that much.

Reviewed By: xavierd

Differential Revision: D15459363

fbshipit-source-id: a022ca5b26138177637ff92a1262e7831e8434d4