indexedlog: reuse indexes to speed up non-racy log write
Summary:
The previous patch made it possible that the in-memory index is consistent with
the desired on-disk index if the on-disk log does not change.

Use that to avoid re-calculating indexes in the write path.

This makes the non-racy write path much faster, while the racy write path is
slow because indexes have to be rebuilt for correctness.

Before:

  log sync (write)                                  102.603 ms   1.760 KB/370.366 MB
  log sync (read 1, write N)                        108.959 ms   1.760 KB/370.364 MB

After:

  log sync (write)                                   29.816 ms     800 B /370.366 MB
  log sync (read 1, write N)                        114.078 ms   1.760 KB/370.364 MB

Reviewed By: xavierd

Differential Revision: D15459366

fbshipit-source-id: caeb328c9fa9e0f9beeebbc7b61c8cc11c1efc80