indexedlog: optimize RotateLog::sync()
Summary:
The so-called "read-only fast path" can actually be slow because it reloads all
logs, just to bypass the directory lock. Optimize it by avoiding that. In the
common read-only path, the logic would be effectively just `writable_log.sync()`,
which is already smart enough to avoid overhead.

This diff can be easier reviewed if all whitespace are ignored.

Reviewed By: xavierd

Differential Revision: D15476301

fbshipit-source-id: 7897d23e2c0b5441efb80f7bf0a3009707dbf95d