indexedlog: add benchmark about lock overhead
Summary:
I was wondering if the additional filesystem lock (ex. RotateLog has a lock,
which makes Log's lock unnecessary, which in turn makes Index's lock
unnecessary) should be removed. Add a benchmark to get some sense about it.

It seems it's pretty fast on POSIX at least:

  lock a directory                                    0.001 ms
  open and lock a directory                           0.005 ms

So probably no short-term plan to skip the unnecessary lock.

Some modules and structs are made public to make the benchmark possible.

Reviewed By: xavierd

Differential Revision: D15481925

fbshipit-source-id: 23c55862e368c08255efa73f6b5514bfed4d7a5a