Fix potential null dereference in HgBackingStore tests
Summary:
HgBackingStore has two constructors. The constructor used in production code accepts an EdenStats object, but the constructor used in test code does not. This means that `HgBackingStore::stats_` could be null, thus tracking statistics in HgBackingStore is more work than it should be.

Make HgBackingStore require an EdenStats object so `HgBackingStore::stats_` is never null.

In production code, this diff should not change behavior.

Reviewed By: simpkins

Differential Revision: D15256266

fbshipit-source-id: f74137bed3e83e1b82d68b840fa55f3a3f0bf0cc