visibility: move to changelog
Summary:
If the visibility state file is read after the changelog is opened, the visibleheads
may have been modified to point to a commit that had been added to the changelog
afterwards, which means the commit is not accessible to the current process.

We need to ensure that the visibleheads file is read just before the changelog is
opened.  The safest way to handle this is to make the visibleheads tracking part
of the changelog object.  Whenever we open the changelog, we get a correctly
up-to-date view of the visible heads.

Reviewed By: quark-zju

Differential Revision: D15468483

fbshipit-source-id: 94d6b3e6db5211d7346f4a244f182383970fcb74